refactor(monument): rename DTO package and update entity fields
- Rename `pt.estga.monument.dots` package to `pt.estga.monument.dtos` and move associated DTO classes
- Update DTO import statements across `monument` and `bookmark` modules to reflect the new package name
- Replace `street` and `houseNumber` fields with `address` and `postalCode` in the `Monument` entity
- Remove `externalId` field from the `Monument` entity
- Remove `findByExternalId` query method from `MonumentRepository`
