+++ bookmark/src/main/java/pt/estga/bookmark/dto/MonumentBookmarkContent.java
@@ -1,6 +1,6 @@
 package pt.estga.bookmark.dto;
 
-import pt.estga.monument.dots.MonumentDto;
+import pt.estga.monument.dtos.MonumentDto;
 
 public record MonumentBookmarkContent(MonumentDto monument) implements BookmarkContent {
 }
