+++ monument/src/main/java/pt/estga/monument/controllers/MonumentAdminController.java
@@ -11,8 +11,8 @@ import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import pt.estga.monument.Monument;
 import pt.estga.monument.MonumentMapper;
 import pt.estga.monument.services.MonumentService;
-import pt.estga.monument.dots.MonumentDto;
-import pt.estga.monument.dots.MonumentRequestDto;
+import pt.estga.monument.dtos.MonumentDto;
+import pt.estga.monument.dtos.MonumentRequestDto;
 import pt.estga.sharedweb.exceptions.ResourceNotFoundException;
 
 import java.net.URI;
